Update App.jsx, Header.jsx, home.jsx and move files to layout directory
