+++ src/components/ui/Header.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import CartIcon from './CartIcon';
-import Nav from '../Nav';
+import Nav from '../../layout/Nav';
 
 const Header = () => {
   // Set the value of isMobileMenuOpen based on your logic
